code/requirements.txt
gensim==3.8.0
goatools==1.0.3
matplotlib==3.1.3
networkx==2.4
numpy==1.18.1
numpy-indexed==0.3.5
pandas==1.0.3
pydot==1.4.1
requests==2.23.0
scikit-learn==0.22.1
scipy==1.4.1
statsmodels==0.11.1
umap-learn==0.4.1
urllib3==1.25.8
wget==3.2
hdbscan==0.8.26
seaborn==0.10.1